+++ pyoxidizer/src/pyrepackager/repackage.rs
@@ -1468,6 +1468,7 @@ pub fn derive_python_config(
          unbuffered_stdio: {},\n    \
          parser_debug: {},\n    \
          quiet: {},\n    \
+         use_hash_seed: {},\n    \
          verbose: {},\n    \
          frozen_importlib_data: include_bytes!(r#\"{}\"#),\n    \
          frozen_importlib_external_data: include_bytes!(r#\"{}\"#),\n    \
@@ -1508,6 +1509,7 @@ pub fn derive_python_config(
         embedded.unbuffered_stdio,
         embedded.parser_debug,
         embedded.quiet,
+        embedded.use_hash_seed,
         embedded.verbose,
         importlib_bootstrap_path.display(),
         importlib_bootstrap_external_path.display(),